Seperate the bytebeat crate into two crates
This commit seperates the crate into two crates. The first one, bytebeat, now just contains encoding and twitter bot controlling code while bytebot-rpn contains the code needed to parse, compile, and evaluate bytebeats.
