+++ bytebot-rpn/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "bytebot-rpn"
+version = "0.1.0"
+authors = ["Aaron Kofsky <a2aarontothe2@gmail.com>", "Caleb Jones <code@calebjones.net>"]
+
+[dependencies]
