+++ src/bin/beat.rs
@@ -1,4 +1,4 @@
-extern crate bytebeat;
+extern crate bytebot_rpn as rpn;
 
 use std::io::{Read, Write};
 
@@ -6,11 +6,11 @@ use std::io::{Read, Write};
 fn main() {
     let mut code = String::new();
     std::io::stdin().read_to_string(&mut code).unwrap();
-    let code = bytebeat::parse_beat(&code).unwrap();
-    let code = bytebeat::compile(code).unwrap();
+    let code = rpn::parse_beat(&code).unwrap();
+    let code = rpn::compile(code).unwrap();
     for i in 0..60 {
         let buf: Vec<_> = (i * 8000..(i + 1) * 8000)
-            .map(|t| bytebeat::eval_beat(&code, t).into())
+            .map(|t| rpn::eval_beat(&code, t).into())
             .collect();
         std::io::stdout().write_all(&buf[..]).unwrap();
     }
