+++ src/bin/random.rs
@@ -1,3 +1,4 @@
+extern crate bytebot_rpn as rpn;
 extern crate bytebeat;
 extern crate rand;
 
@@ -10,13 +11,13 @@ fn main() {
     println!("{}", code);
 }
 
-fn gen_beat(min: usize, max: usize) -> bytebeat::Program {
+fn gen_beat(min: usize, max: usize) -> rpn::Program {
     loop {
         let length = rand::thread_rng().gen_range::<usize>(min, max);
         let code = random::random_t_multiply(length);
-        let code = bytebeat::compile(code).unwrap();
+        let code = rpn::compile(code).unwrap();
         let buf: Vec<_> = (0..5 * 8000)
-            .map(|t| bytebeat::eval_beat(&code, t))
+            .map(|t| rpn::eval_beat(&code, t))
             .collect();
 
         let is_silent = buf.iter().all(|&x| x == buf[0]);
